+++ src/hellonet.h
@@ -14,10 +14,10 @@
 class HelloNet{
 private:
     //array of weighs
-    std::vector<std::vector<std::vector<float>>>  weights;
+    std::vector<std::vector<std::vector<long double>>>  weights;
 
     //array of bias vectors
-    std::vector<std::vector<float>> biases;
+    std::vector<std::vector<long double>> biases;
 
     //the number of layers in the neural net
     unsigned long num_layers;
@@ -31,40 +31,40 @@ public:
     explicit HelloNet(std::vector<unsigned long> layerConfig);
 
     //constructor for debugging
-    HelloNet(std::vector<unsigned long> layer_config, float fixed_weight, float fixed_bias);
+    HelloNet(std::vector<unsigned long> layer_config, long double fixed_weight, long double fixed_bias);
 
     //print the weight tables to stdout (DEBUG)
     void dumpWeightTables();
 
     //activation function applied to hypothesis h
-    float activate(float h);
+    long double activate(long double h);
 
     //derivative of activation function
-    float actPrime(float h);
+    long double actPrime(long double h);
 
     //classify input data and return result into the same variable passed in
-    void forwardProp(std::vector<float> &data);
+    void forwardProp(std::vector<long double> &data);
 
     //perform back propagation algorithm on a single training sample
-    void backProp(std::vector<float> &trainingLabel,
-                  std::vector<float> &trainingData,
-                  std::vector<std::vector<float>> &nablaB,
-                  std::vector<std::vector<std::vector<float>>> &nablaW);
+    void backProp(std::vector<long double> &trainingLabel,
+                  std::vector<long double> &trainingData,
+                  std::vector<std::vector<long double>> &nablaB,
+                  std::vector<std::vector<std::vector<long double>>> &nablaW);
 
     //get the d/dC by performing a piecewise subtraction between two vectors and gets loaded into output á la C-style
-    void costDerivative(std::vector<float> &expectedValues,
-                        std::vector<float> &currentValues,
-                        std::vector<float> &output);
+    void costDerivative(std::vector<long double> &expectedValues,
+                        std::vector<long double> &currentValues,
+                        std::vector<long double> &output);
 
     //train the NN using stochastic gradient descent
     void sgd(unsigned long epochs,
-             float learnRate,
-             std::vector<std::vector<float>> &trainingData,
-             std::vector<std::vector<float>> &labels);
+             long double learnRate,
+             std::vector<std::vector<long double>> &trainingData,
+             std::vector<std::vector<long double>> &labels);
 
     //here's an implementation of simple gradient descent
-    void gradientDescend(float learnRate,
-                         std::vector<std::vector<float>> &trainingData,
-                         std::vector<std::vector<float>> &labels);
+    void gradientDescend(long double learnRate,
+                         std::vector<std::vector<long double>> &trainingData,
+                         std::vector<std::vector<long double>> &labels);
 
 };
